+++ iam.tf
@@ -1,0 +1,11 @@
+resource "google_service_account" "nomad" {
+  account_id   = "nomad-sa"
+  display_name = "Custom Nomad and Consul service account"
+}
+
+# https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/google_project_iam#google_project_iam_member
+resource "google_project_iam_member" "nomad" {
+  project = var.project_id
+  role    = "roles/compute.viewer"
+  member  = "serviceAccount:${google_service_account.nomad.email}"
+}